moved some things from App to ThoughtsFeed component
