+++ code/src/components/LikeButton/LikeButton.module.css
@@ -3,7 +3,7 @@
   border: none;
   border-radius: 50%;
   font-size: 1rem;
-  padding: 0.7rem;
+  padding: 0.7rem 0.9rem;
 }
 
 .likedlikeButton {
@@ -11,5 +11,5 @@
   border: none;
   border-radius: 50%;
   font-size: 1rem;
-  padding: 0.7rem;
+  padding: 0.7rem 0.9rem;
 }
