+++ code/src/components/ThoughtsFeed/ThoughtsFeed.jsx
@@ -1,11 +1,28 @@
 /* eslint-disable no-underscore-dangle */
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 import moment from 'moment';
 
+import ThoughtForm from 'components/ThoughtForm/ThoughtForm';
 import LikeButton from '../LikeButton/LikeButton';
 import styles from './ThoughtsFeed.module.css';
 
-const ThoughtsFeed = ({ loading, thoughtsFeed, setThoughtsFeed }) => {
+const ThoughtsFeed = () => {
+  const [thoughtsFeed, setThoughtsFeed] = useState([]);
+  const [loading, setLoading] = useState(false);
+
+  const fetchThoughts = () => {
+    setLoading(true);
+    fetch('https://happy-thoughts-technigo.herokuapp.com/thoughts')
+      .then((res) => res.json())
+      .then((data) => setThoughtsFeed(data))
+      .catch((error) => console.error(error))
+      .finally(() => setLoading(false));
+  };
+
+  useEffect(() => {
+    fetchThoughts();
+  }, []);
+
   if (loading) {
     return <h2 className={styles.loadingMessage}>Loading...</h2>
   }
@@ -22,6 +39,7 @@ const ThoughtsFeed = ({ loading, thoughtsFeed, setThoughtsFeed }) => {
 
   return (
     <section className={styles.feedGrid}>
+      <ThoughtForm setThoughtsFeed={setThoughtsFeed} />
       {thoughtsFeed.map((thought) => (
         <div key={thought._id} className={styles.thoughtCard}>
           <h4 className={styles.thoughtMessage}>{thought.message}</h4>
