+++ code/src/components/ThoughtsFeed/ThoughtsFeed.module.css
@@ -1,12 +1,9 @@
-.loadingMessage {
-}
-
 .feedGrid {
   background-color: whitesmoke;
   display: grid;
   grid-template-columns: 1fr;
+  max-width: 700px;
   gap: 2rem;
-  width: 100%;
 }
 
 .thoughtCard {
@@ -42,4 +39,13 @@
 }
 
 .timePosted {
+}
+
+.loadingMessage {
+}
+
+@media (max-width: 705px) {
+  .feedGrid {
+    width: 90%;
+  }
 }
